+++ app/core/BackupVault/index.ts
@@ -1,0 +1,5 @@
+export {
+  backupVault,
+  getVaultFromBackup,
+  resetVaultBackup,
+} from './backupVault';
